… libvips/include/vips/vips7compat.h