Hi John,

Replacing the "Duff's device" with a for-loop seems to allow gcc to optimise this code, including permitting the use of vectorisation. It reduces the CPU instruction count of `conv_gen` by about 30%, which leads to a performance improvement to the (non-3x3) convolution operation as a whole by somewhere between 20 to 25%.

Hopefully I've not made any glaring errors in my maths this time!

The following tests should demonstrate the fairly dramatic drop in instruction count for a Gaussian blur with sigma=10 and liborc disabled. (I've included libjpeg's `decode_mcu` function in the output as a "control" to ensure the same test was run for each scenario.)

Cheers,
Lovell


precision=integer
```
-- master -O2

4.415s

18,956,604,353  libvips/convolution/im_conv.c:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
15,287,450,934  libvips/convolution/im_conv.c:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- master -O3

4.432s

18,956,603,353  ???:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
15,287,457,226  ???:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- conv-allow-vectorise -O2

4.190s

13,974,480,980  libvips/convolution/im_conv.c:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
11,269,633,462  libvips/convolution/im_conv.c:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- conv-allow-vectorise -O3

3.395s

11,362,912,900  ???:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
 9,163,516,484  ???:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]
```
precision=float
```
-- master -O2

4.525s

18,305,707,960  libvips/convolution/im_conv_f.c:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
13,264,259,684  libvips/convolution/im_conv_f.c:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- master -O3

4.439s

20,164,137,120  ???:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
13,264,371,184  ???:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- conv-allow-vectorise -O2

4.210s

17,550,623,384  libvips/convolution/im_conv_f.c:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
12,655,099,920  libvips/convolution/im_conv_f.c:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]

-- conv-allow-vectorise -O3

3.926s

17,891,636,579  ???:conv_gen'2 [/usr/local/lib/libvips.so.42.3.1]
11,593,399,095  ???:conv_gen [/usr/local/lib/libvips.so.42.3.1]
   592,029,158  ???:decode_mcu [/usr/local/lib/libjpeg.so.8.1.2]
```