@jcupitt 2016-01-26
Wow Lovell, that looks fantastic. I've not reworked im_conv() for years :-( certainly not since gcc auto-vec became useful. Duff was a good thing back then, but it's obviously not now. Brilliant!