@jcupitt 2016-01-26
I ran the vector analyzer and it's not actually being vectorised. It doesn't seem to like the `p[i][x]` in the inner loop -- we'd need to find another way to code this.

The error is:

```
im_conv.c:773:3: note: bad data references.
```

I made a branch here to mess about:

https://github.com/jcupitt/libvips/tree/convolution-tuning

The uchar loop for `im_conv.c` is there three times: as duff, as not-duff, and as expanded not-duff (useful for checking loop line numbers with the vec tree dump).

I do see a very nice speedup with your change though, I guess gcc is able to do something clever even without vectorising:

First make a 4x4 sharpen mask (so we avoid the special 3x3 path):

```
$ cat >sharp.mat
4 4  
-1 -1 -1 -1
-1 3 3 -1
-1 3 3 -1
-1 -1 -1 -1
```

Then `conv` as was:

```
$ time vips conv wtc.tif x.tif sharp.mat --vips-novector
real	0m4.974s
user	0m15.992s
sys	0m0.376s
```

Your duff-free conv:

```
$ time vips conv wtc.tif x.tif sharp.mat --vips-novector
real	0m3.414s
user	0m10.092s
sys	0m0.396s
```

And finally Orc:

```
$ time vips conv wtc.tif x.tif sharp.mat 
real	0m1.368s
user	0m2.268s
sys	0m0.384s
```

Which is rather sobering.