@lovell 2016-01-26
Thanks for merging John.

My attempt to break the inner loop apart, from:
```c
for ( i = 0; i < nnz; i++ )
  sum += t[i] * p[i][x];
```
to:
```c
int t2[nnz];
for ( i = 0; i < nnz; i++ )
  t2[i] = p[i][x];
for ( i = 0; i < nnz; i++ )
  sum += t[i] * t2[i];
```
did vectorise but with worse performance :(

Are the under/overflow counts in the `VIPS_CLIP_` macros used by anything? If not, a possible improvement might be using the compiler's built-in min and max functions for pixel value clipping (with the aim to avoid branching).