@jcupitt 2016-01-26
I tried:

```C
                for ( i = 0; i < nnz; i++ ) {
                        sum += t[i] * *p[i];
                        p[i] += 1; 
                }
```

But it's hardly different and also didn't vectorise :( 

At the moment, conv tries to squeeze out zero mask elements. If we abandon that we could remove a level of indirection in the accesses to pixels and make it vectorisable (I think). I'll see if I can make a test.

Yes, it looks like GCC has __builtin_fmin()/_fmax(), but nothing for int values. We should probably make a VIPS_FMIN()/_FMAX() and use it where we can. 

I think g++ has a min/max for int with the built in `a <? b`.