@jcupitt 2016-01-26
Yes, swapping the

```C
IM_CLIP_UCHAR( sum, seq );
```

for 

```C
sum = VIPS_CLIP( 0, sum, 255 );
```

Saves about 10%:

```
$ time vips conv wtc.tif x.tif sharp.mat --vips-novector
real	0m3.137s
user	0m9.008s
sys	0m0.440s
```

`__builtin_fmaxf` is actually slower, since it's having to go to float and back:

```C
sum = __builtin_fminf( 255.0f, __builtin_fmaxf( 0.0f, sum ) );
```

```
$ time vips conv wtc.tif x.tif sharp.mat --vips-novector
real	0m2.761s
user	0m9.472s
sys	0m0.404s
```