@lovell 2016-01-26
To use fast integer-based min/max clipping on all compilers (not just gcc), we'd need to use SSE2 intrinsics, available on all 64-bit Intel CPUs.

This would look something like:
```c
#if defined(__x86_64__)
#  include <emmintrin.h>
#  define VIPS_CLIP_INT( A, V, B ) _mm_extract_epi16( _mm_min_epi16( _mm_max_epi16( _mm_set1_epi16( V ), _mm_set1_epi16( A ) ), _mm_set1_epi16( B ) ), 0 )
#else
#  define VIPS_CLIP_INT( A, V, B ) VIPS_CLIP( A, V, B )
#endif
```
The real power here is the ability to clip 8 integers at once (with `_mm_set_epi16` rather than `_mm_set1_epi16`), but that's a world of complexity I'd be happy to ignore by hiding behind liborc :)