@jcupitt 2016-01-26
I tried adding `VIPS_ICLIP()` with that scrap of sse, but it's no faster than `?:`, unfortunately.

I timed the 3x3 path and it still seems usefully quicker.