@jcupitt 2016-01-26
Shall we stop here? The easy stuff seems to be done. 

I'll make 8.2.2 with all these changes.