@lovell 2016-01-26
I agree, release at will, thanks John!